requirements.txt

# --- Core Django stack ---
Django==4.2.16
channels==4.1.0
daphne==4.0.0
asgiref==3.8.1               # Phiên bản phù hợp với Django 4.2

# --- Django extensions ---
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-jsonapi==7.1.0

# --- Environment ---
python-dotenv==1.0.1

# --- ML & CV ---
gunicorn>=21.2


opencv-python-headless==4.10.0.84
numpy==1.26.4   # hoặc thấp hơn, có wheel sẵn
pillow==10.2.0
matplotlib==3.8.4
sympy>=1.13.3,<2.0


# --- Networking & Security ---
httpx==0.28.1
cryptography==42.0.7
pyOpenSSL==24.1.0
service-identity==24.1.0
Twisted==23.10.0
autobahn==23.6.2
txaio==23.1.1
attrs==23.2.0
zope.interface==6.1
incremental==24.7.2 
# --- Utilities ---
packaging==24.1
python-dateutil==2.9.0.post0
sqlparse==0.5.1
typing_extensions==4.12.2
tomli==2.0.1


anyio==4.11.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
huggingface_hub==1.1.2
idna==3.11
PyYAML==6.0.3
shellingham==1.5.4
sniffio==1.3.1
tqdm==4.67.1
typer-slim==0.20.0
tensorflow==2.20.0

django-cors-headers==4.9.0
